Unify bridge session turn inference
