+++ rust/crates/runtime/src/server/mod.rs
@@ -55,6 +55,7 @@ pub mod server_loop_host;
 pub mod server_skill_subrun;
 pub mod server_tool_executor;
 pub(crate) mod session;
+pub(crate) mod session_turn;
 mod skillify_agent_executor;
 mod state_builder;
 pub mod sweeper_lease;
