+++ rust/crates/runtime/src/server/run/lifecycle/persistence.rs
@@ -680,25 +680,6 @@ async fn persist_trace_degraded_event(
     }
 }
 
-pub(crate) async fn infer_session_turn(
-    shared_pool: Option<&SharedPool>,
-    user_id: &str,
-    session_id: &str,
-) -> u32 {
-    let Some(shared_pool) = shared_pool else {
-        return 1;
-    };
-    let count: i64 = sqlx::query_scalar(
-        "SELECT COUNT(*) FROM agent_events WHERE user_id = ? AND session_id = ? AND event_type = 'user_query'",
-    )
-    .bind(user_id)
-    .bind(session_id)
-    .fetch_one(shared_pool.get())
-    .await
-    .unwrap_or(0);
-    (count.max(0) as u32).saturating_add(1)
-}
-
 /// Persist `user_query` + `llm_response` core events to `agent_events` after
 /// the server-driven agentic loop completes.  This closes the persistence gap
 /// where the bridge path (`/chat/turn`) wrote these events but the server loop
