+++ rust/crates/runtime/src/turn/bridge/inprocess.rs
@@ -1068,25 +1068,6 @@ fn turn_count_from_messages(messages: &[Value]) -> i64 {
         .count() as i64
 }
 
-async fn infer_bridge_session_turn(
-    shared_pool: Option<&SharedPool>,
-    user_id: &str,
-    session_id: &str,
-) -> u32 {
-    let Some(shared_pool) = shared_pool else {
-        return 1;
-    };
-    let count: i64 = sqlx::query_scalar(
-        "SELECT COUNT(*) FROM agent_events WHERE user_id = ? AND session_id = ? AND event_type = 'user_query'",
-    )
-    .bind(user_id)
-    .bind(session_id)
-    .fetch_one(shared_pool.get())
-    .await
-    .unwrap_or(0);
-    (count.max(0) as u32).saturating_add(1)
-}
-
 fn bridge_root_turn_journal_owned(
     headers: &HeaderMap,
     payload: &Value,
@@ -1366,7 +1347,12 @@ impl InProcessChatTurnBridge {
             (turn, "header")
         } else if !session_id.is_empty() {
             (
-                infer_bridge_session_turn(shared_pool.as_ref(), &user_id, &session_id).await,
+                crate::server::session_turn::infer_session_turn(
+                    shared_pool.as_ref(),
+                    &user_id,
+                    &session_id,
+                )
+                .await,
                 "inferred_agent_events",
             )
         } else {
